feat: Add feature flag for taken/live
To allow us to push this to production, without going live, this
commit adds a flag to the configuration object for setting the live
status of the app. If the app is not live, the /zaken/ routes will
return http 404.

We also add a flag for the 'taken'-component. This is experimental,
and should not be live yet, even when the app itself can go live.