feat: kvk zaken (#302)
* feat: Add user types
Prepare for using organisations and persons. We add an abstracted
User interface, which is implemented by both a Person and an
organisation class.

* feat: OpenZaakClient now takes baseUrl as string
The OpenZaakClient now accepts both strings and URLs for the baseURL.
Tests rely on baseURL being an empty string, which is not a valid URL.
Since the URL is stringified anyway in axios, and we do pass the env
variable as a URL, no url safety in production is lost.

* feat: Zaken can now handle organisations
Zaken now instead of receiving a bsn, will receive a 'user' object.
This can be a person or an organisation. For organisations, we have
to gather 'zaken' differently, using the 'rol' endpoint first.