chore: Refactor Zaken to provide user with request
To prepare for reuse/caching of metadata, each request for Zaken
now requires providing the user for this request, instead of keeping
this with the constructor.

This hopefully allows for moving construction of the Zaken-object to
the lambda initializer, which in turn will allow for pre-loading
metadat from the ZGW-api (zaaktypen, zaaktypecatalogi, etc.) which are
user-independent.